requirements.txt
numpy~=1.18.5
pandas~=1.0.4
pytest~=5.4.3
pygame~=1.9.6
Pillow~=7.1.2